requirements.txt
opencv-python == 4.8.1.78
onnxruntime == 1.16.3
numpy == 1.26.2
Pillow == 10.1.0

